refactor: Controller 테스트 분리
